requirements.txt
streamlit==1.44.1
numpy>=1.23,<3
scikit-learn==1.0.2
joblib==1.2.0
xgboost==1.6.2
Pillow==9.5.0
setuptools==59.8.0